Rename the crate to "enum_match". Remove struct generation
Unfortunately, struct generation confuses rust-analyzer to the point
where writing Rust code is painful. This applies not to writing the
enum_gen-enums, but to any code that uses those enums.

For this reason, restrict the crate to just generating match
expressions. The structs now need to be defined by the user.

This commit updates the code documentation and the unit test, but
not all of the intended functionality is implemented yet. Doctests
currently fail. This will be fixed in subsequent patches.